requirements.txt
Jinja2~=2.10.0
PyYAML~=3.13.0
numpy~=1.15.0
